requirements.txt
flask
gunicorn
snowflake-connector-python